// src/types/coinDetails.ts
export type CoinDetails = {
  "24hVolume": string;
  allTimeHigh: {
    price: string;
    timestamp: number;
  };
  change: string;
  iconUrl: string;
  listedAt: number;
  marketCap: string;
  name: string;
  numberOfExchanges: number;
  numberOfMarkets: number;
  price: string;
  priceAt: number;
  rank: number;
  symbol: string;
  websiteUrl: string;
};